+++ src/rex.mart/NEWS.rst
@@ -5,7 +5,7 @@ REX.MART Change Log
 .. contents:: Table of Contents
 
 
-0.2.0 (2016-xx-xx)
+0.2.0 (2016-03-23)
 ==================
 
 - Added ability to specify additional HTSQL extensions to enable in the